requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
opensearch-py==2.6.0
neo4j==5.23.1
oracledb==2.3.0
python-dotenv==1.0.1
pydantic==2.8.2
strands-agents>=0.1.0
